Switch TX to ISR and aligned addressing

+++ USB-Drivers/SimpleCDC.c
@@ -142,7 +142,8 @@ uint8_t CDC_Device_SendByte_Prep(USB_ClassInfo_CDC_Device_t* const CDCInterfaceI
 }
 
 
-uint16_t CDC_Device_BytesReceived(USB_ClassInfo_CDC_Device_t* const CDCInterfaceInfo)
+
+uint8_t CDC_Device_BytesReceived(USB_ClassInfo_CDC_Device_t* const CDCInterfaceInfo)
 {
 	if (!(CDCInterfaceInfo->State.LineEncoding.BaudRateBPS))
 	  return 0;
@@ -166,6 +167,9 @@ uint16_t CDC_Device_BytesReceived(USB_ClassInfo_CDC_Device_t* const CDCInterface
 		return 0;
 	}
 }
+
+
+
 
 int16_t CDC_Device_ReceiveByte(USB_ClassInfo_CDC_Device_t* const CDCInterfaceInfo)
 {
